renamed fluidsynth.[ch] to music_fluidsynth.[ch]
fixed a case where the compiler got confused about which fluidsynth.h to
include...
